Move GlobalGraphID and TinyTimer out of utils [WIP]
